* dispextern.h (FACE_SUITABLE_FOR_CHAR_P): Remove unused macro. (FACE_FOR_CHAR): Simplify because face_for_char does the same. * fontset.c (face_suitable_for_char_p) [0]: Remove unused function. (font_for_char): Prefer ptrdiff_t to int for buffer position. (face_for_char): Likewise.  Rearrange eassert and return ASCII face for CHAR_BYTE8_P. * fontset.h (font_for_char, face_for_char): Adjust prototypes.

+++ src/fontset.h
@@ -28,8 +28,8 @@ along with GNU Emacs.  If not, see <http://www.gnu.org/licenses/>.  */
 
 extern void free_face_fontset (struct frame *, struct face *);
 extern int face_for_char (struct frame *, struct face *, int,
-                          int, Lisp_Object);
-extern Lisp_Object font_for_char (struct face *, int, int, Lisp_Object);
+                          ptrdiff_t, Lisp_Object);
+extern Lisp_Object font_for_char (struct face *, int, ptrdiff_t, Lisp_Object);
 
 extern int make_fontset_for_ascii_face (struct frame *, int, struct face *);
 extern int fontset_from_font (Lisp_Object);
